refactor: add modal과 edit modal을 manage modal로 통일
