+++ frontend/src/components/ChannelList/ChannelListHeader/ChannelListHeader.tsx
@@ -2,20 +2,20 @@ import { Button } from '@mui/material';
 import { Box } from '@mui/system';
 import { useState } from 'react';
 
-import ChannelCreateModal from './ChannelCreateModal';
+import ChannelManageModal from './ChannelManageModal';
 
 export default function ChannelListHeader(): JSX.Element {
   const [open, setOpen] = useState(false);
-  const handleOpen = (): void => {
-    setOpen(true);
-  };
-  const handleClose = (): void => {
-    setOpen(false);
-  };
 
   return (
     <Box display="flex" alignItems="center">
-      <Button variant="contained" sx={{ m: 1, p: 2 }} onClick={handleOpen}>
+      <Button
+        variant="contained"
+        sx={{ m: 1, p: 2 }}
+        onClick={() => {
+          setOpen(true);
+        }}
+      >
         Create a new channel
       </Button>
       {/* <Button variant="contained" color="secondary" sx={{ m: 1, p: 2 }}>
@@ -33,7 +33,7 @@ export default function ChannelListHeader(): JSX.Element {
           />
         </FormGroup>
       </Box> */}
-      <ChannelCreateModal {...{ open, handleClose }} />
+      <ChannelManageModal {...{ open, setOpen }} isAddChannel={true} />
     </Box>
   );
 }
